101-print_number.c

+++ more_functions_nested_loops/101-print_number.c
@@ -8,15 +8,15 @@
  */
 void print_number(int n)
 {
-	unsigned int num;
+	int num;
 
 	if (n < 0)
 	{
 		_putchar('-');
-		num = -((unsigned int)n);
+		num = -(n);
 	}
 	else
-		num = (unsigned int) n;
+		num = n;
 
 	if (num / 10)
 		print_number(num / 10);
